+++ stroom-core-shared/src/main/java/stroom/streamtask/shared/FetchStreamTaskSummaryAction.java
@@ -16,29 +16,18 @@
 
 package stroom.streamtask.shared;
 
-import stroom.entity.shared.Action;
-import stroom.entity.shared.ResultList;
-import stroom.query.api.v2.ExpressionOperator;
+import stroom.entity.shared.ExpressionCriteria;
+import stroom.entity.shared.FindAction;
 
-public class FetchStreamTaskSummaryAction extends Action<ResultList<StreamTaskSummary>> {
+public class FetchStreamTaskSummaryAction extends FindAction<ExpressionCriteria, StreamTaskSummary> {
     private static final long serialVersionUID = -1773544031158236156L;
-
-    private ExpressionOperator expression;
 
     public FetchStreamTaskSummaryAction() {
         // Default constructor necessary for GWT serialisation.
     }
 
-    public FetchStreamTaskSummaryAction(final ExpressionOperator expression) {
-        this.expression = expression;
-    }
-
-    public ExpressionOperator getExpression() {
-        return expression;
-    }
-
-    public void setExpression(final ExpressionOperator expression) {
-        this.expression = expression;
+    public FetchStreamTaskSummaryAction(final ExpressionCriteria criteria) {
+        super(criteria);
     }
 
     @Override